+++ ChangeLog.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `DataApi::request*`: default `$requestPacket` to `[]` rather than `null`,
+    which would cause cryptic errors if no packet is specified.
 
 ## [v0.10.0] - 2018-09-17
 ### Added