+++ ChangeLog.md
@@ -9,7 +9,8 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 ### Fixed
 - `DataApi::request*`: default `$requestPacket` to `[]` rather than `null`,
-    which would cause cryptic errors if no packet is specified.
+    which would cause cryptic errors if no packet is specified. Additionally, a
+    warning is provided if the `$requestPacket` is not a PHP array.
 
 ## [v0.10.0] - 2018-09-17
 ### Added